refactor(layour, Footer): remove unused, styling
